+++ pkg/application/app.go
@@ -30,10 +30,12 @@ type Application struct {
 	Catalog              string
 	Values               string
 	InCluster            bool
+	ClusterName          string
 	Organization         organization.Org
 	UserConfigSecretName string
 	ExtraConfigs         []applicationv1alpha1.AppExtraConfig
 	RepoName             string
+	InstallNamespace     string
 
 	AppLabels       map[string]string
 	ConfigMapLabels map[string]string
@@ -42,14 +44,16 @@ type Application struct {
 // New creates a new Application
 func New(installName string, appName string) *Application {
 	return &Application{
-		InstallName:  installName,
-		AppName:      appName,
-		RepoName:     appName,
-		Version:      "",
-		Catalog:      "cluster",
-		Values:       "\n",
-		InCluster:    true,
-		Organization: *organization.New("giantswarm"),
+		InstallName:      installName,
+		AppName:          appName,
+		RepoName:         appName,
+		ClusterName:      "",
+		Version:          "",
+		Catalog:          "cluster",
+		Values:           "\n",
+		InCluster:        true,
+		Organization:     *organization.New("giantswarm"),
+		InstallNamespace: "org-giantswarm",
 	}
 }
 
@@ -172,6 +176,20 @@ func (a *Application) WithRepoName(repoName string) *Application {
 	return a
 }
 
+// WithClusterName sets the name of the cluster the app with be installed into.
+// This is used for populating the appropriate labels on the App resources.
+func (a *Application) WithClusterName(clusterName string) *Application {
+	a.ClusterName = clusterName
+	return a
+}
+
+// WithInstallNamespace sets the namespace used by helm to install the chart
+// This can be different to the namespace the App CR is in.
+func (a *Application) WithInstallNamespace(namespace string) *Application {
+	a.InstallNamespace = namespace
+	return a
+}
+
 // Build generates the App and ConfigMap resources
 func (a *Application) Build() (*applicationv1alpha1.App, *corev1.ConfigMap, error) {
 	switch a.Version {
@@ -196,12 +214,23 @@ func (a *Application) Build() (*applicationv1alpha1.App, *corev1.ConfigMap, erro
 		a = a.WithVersion(latestVersion)
 	}
 
+	if !a.InCluster && a.ClusterName == "" {
+		return nil, nil, fmt.Errorf("a `ClusterName` must be provided when `InCluster` is set to `false`")
+	}
+
+	// Use the install namespace if provided, otherwise default to the Orgs namespace
+	namespace := a.InstallNamespace
+	if namespace == "" {
+		namespace = a.Organization.GetNamespace()
+	}
+
 	appTemplate, err := templateapp.NewAppCR(templateapp.Config{
 		AppName:                 a.InstallName,
 		Name:                    a.AppName,
 		Catalog:                 a.Catalog,
 		InCluster:               a.InCluster,
-		Namespace:               a.Organization.GetNamespace(),
+		Cluster:                 a.ClusterName,
+		Namespace:               namespace,
 		Organization:            a.Organization.Name,
 		UserConfigConfigMapName: fmt.Sprintf("%s-userconfig", a.InstallName),
 		UserConfigSecretName:    a.UserConfigSecretName,
@@ -221,6 +250,11 @@ func (a *Application) Build() (*applicationv1alpha1.App, *corev1.ConfigMap, erro
 	}
 	if a.AppLabels != nil {
 		app.SetLabels(a.AppLabels)
+	}
+
+	if !app.Spec.KubeConfig.InCluster {
+		// We need to fix the kubeconfig context name to match the format that is used by CAPI
+		app.Spec.KubeConfig.Context.Name = fmt.Sprintf("%s-admin@%s", a.ClusterName, a.ClusterName)
 	}
 
 	configmap, err := templateapp.NewConfigMap(templateapp.UserConfig{