+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `WithNamespace` and `WithClusterName` to Application to support installing apps into workload clusters
+- Added `WithInstallNamespace` and `WithClusterName` to Application to support installing apps into workload clusters
 - Added error handler to ensure a `ClusterName` is provided with an Application if `InCluster` is set to `false`
 
 ### Fixed