+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `WithInstallNamespace` and `WithClusterName` to Application to support installing apps into workload clusters
+- Added error handler to ensure a `ClusterName` is provided with an Application if `InCluster` is set to `false`
+
+### Fixed
+
+- Correctly set the `app.Spec.KubeConfig.Context.Name` to the value used by CAPI
+
 ## [0.5.0] - 2023-09-14
 
 ### Added