+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2023-09-14
+
 ### Added
 
 - Added `WithInstallNamespace` and `WithClusterName` to Application to support installing apps into workload clusters
@@ -200,7 +202,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/giantswarm/clustertest/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/giantswarm/clustertest/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/giantswarm/clustertest/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/giantswarm/clustertest/compare/v0.3.0...v0.3.1